+++ src/dropdowns/DropdownObj.jsx
@@ -17,6 +17,8 @@
 * 2025/12/18   ITA   1.06     Renamed to from Dropdown2 to DropdownObj, and performed further tweaks and tests towards preparing this component
                               for npm publishing.
 * 2025/12/26   ITA   1.07     Changed the arrow symbols to + and - for better visibility across different platforms.
+* 2025/12/29   ITA   1.08     Placeholder must show the name of the data as provided by the label attribute.
+
 */
 import PropTypes from 'prop-types';
 import { useEffect, useState } from 'react';
@@ -138,7 +140,7 @@ export function DropdownObj({label, // label with which to describe the dropdown
                         type='text' id='searchDropDown' name='searchDropDown' autoComplete='off'
                         aria-label={`Type to Search for ${label}`} aria-required={true} onChange={e=> handleSearch(e)}
                         disabled={isDisabled}
-                        placeholder='Type to search' value={searchText} />
+                        placeholder={`Type to Search for ${label}`} value={searchText} />
                 
                 <div className='dropdown-js-arrow-container dropdown-js-padding dropdown-js-rounded' onClick={e=> toggleShowList(e)}>
                     <span className='dropdown-js-arrow dropdown-js-padding'>{!showItems? "+" : "-"}</span>